+++ apps/frontend/src/app/Components/Character/StatDisplayComponent.tsx
@@ -5,6 +5,7 @@ import {
 } from '@genshin-optimizer/common/ui'
 import { objMap } from '@genshin-optimizer/common/util'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
+import type { NodeDisplay } from '@genshin-optimizer/gi/ui'
 import type { DisplaySub } from '@genshin-optimizer/gi/wr'
 import { customRead } from '@genshin-optimizer/gi/wr'
 import type { MasonryProps } from '@mui/lab'
@@ -14,7 +15,6 @@ import { useContext, useMemo } from 'react'
 import { DataContext } from '../../Context/DataContext'
 import { OptimizationTargetContext } from '../../Context/OptimizationTargetContext'
 import { getDisplayHeader, getDisplaySections } from '../../Formula/DisplayUtil'
-import type { NodeDisplay } from '../../Formula/uiData'
 import CardDark from '../Card/CardDark'
 import { FieldDisplayList, NodeFieldDisplay } from '../FieldDisplay'
