+++ apps/frontend/src/app/Components/FieldDisplay.tsx
@@ -6,6 +6,8 @@ import {
 import { evalIfFunc, valueString } from '@genshin-optimizer/common/util'
 import type { AmpReactionKey } from '@genshin-optimizer/gi/consts'
 import { allAmpReactionKeys } from '@genshin-optimizer/gi/consts'
+import type { NodeDisplay } from '@genshin-optimizer/gi/ui'
+import { nodeVStr, resolveInfo } from '@genshin-optimizer/gi/ui'
 import GroupsIcon from '@mui/icons-material/Groups'
 import HelpIcon from '@mui/icons-material/Help'
 import type { ListProps, PaletteColor } from '@mui/material'
@@ -22,8 +24,6 @@ import type { ReactNode } from 'react'
 import React, { Suspense, useCallback, useContext, useMemo } from 'react'
 import { DataContext } from '../Context/DataContext'
 import { FormulaDataContext } from '../Context/FormulaDataContext'
-import type { NodeDisplay } from '../Formula/api'
-import { nodeVStr, resolveInfo } from '../Formula/uiData'
 import type { IBasicFieldDisplay, IFieldDisplay } from '../Types/fieldDisplay'
 import AmpReactionModeText from './AmpReactionModeText'
