+++ apps/frontend/src/app/Data/Artifacts/dataUtil.tsx
@@ -1,6 +1,10 @@
 import type { ArtifactSetKey } from '@genshin-optimizer/gi/consts'
-import type { Data, DisplaySub } from '@genshin-optimizer/gi/wr'
-import { inferInfoMut, mergeData } from '../../Formula/api'
+import {
+  inferInfoMut,
+  mergeData,
+  type Data,
+  type DisplaySub,
+} from '@genshin-optimizer/gi/wr'
 
 export function dataObjForArtifactSheet(
   key: ArtifactSetKey,
