+++ apps/frontend/src/app/Data/Characters/ShikanoinHeizou/index.tsx
@@ -15,7 +15,7 @@ import {
   target,
   unequal,
 } from '@genshin-optimizer/gi/wr'
-import { absorbableEle } from '../../../Types/consts'
+import { absorbableEle } from '@genshin-optimizer/gi/consts'
 import { cond, st, stg } from '../../SheetUtil'
 import CharacterSheet from '../CharacterSheet'
 import type { ICharacterSheet } from '../ICharacterSheet.d'
