+++ apps/frontend/src/app/PageTeam/CharacterDisplay/CustomMultiTarget/index.tsx
@@ -4,6 +4,7 @@ import { deepClone } from '@genshin-optimizer/common/util'
 import type { CustomMultiTarget } from '@genshin-optimizer/gi/db'
 import { initCustomMultiTarget } from '@genshin-optimizer/gi/db'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
+import { UIData } from '@genshin-optimizer/gi/ui'
 import AddIcon from '@mui/icons-material/Add'
 import CloseIcon from '@mui/icons-material/Close'
 import DashboardCustomizeIcon from '@mui/icons-material/DashboardCustomize'
@@ -33,7 +34,6 @@ import InfoTooltip from '../../../Components/InfoTooltip'
 import ModalWrapper from '../../../Components/ModalWrapper'
 import { DataContext } from '../../../Context/DataContext'
 import { TeamCharacterContext } from '../../../Context/TeamCharacterContext'
-import { UIData } from '../../../Formula/uiData'
 import CustomMultiTargetCard from './CustomMultiTargetCard'
 import CustomMultiTargetImportBtn from './CustomMultiTargetImportBtn'
