+++ apps/frontend/src/app/PageTeam/CharacterDisplay/StatModal.tsx
@@ -2,6 +2,7 @@ import { ColorText } from '@genshin-optimizer/common/ui'
 import type { TeamCharacter } from '@genshin-optimizer/gi/db'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
 import { allEleDmgKeys, allEleResKeys } from '@genshin-optimizer/gi/keymap'
+import { nodeVStr, resolveInfo } from '@genshin-optimizer/gi/ui'
 import type { ReadNode } from '@genshin-optimizer/gi/wr'
 import { allInputPremodKeys, uiInput as input } from '@genshin-optimizer/gi/wr'
 import BarChartIcon from '@mui/icons-material/BarChart'
@@ -30,7 +31,6 @@ import StatEditorList from '../../Components/StatEditorList'
 import { CharacterContext } from '../../Context/CharacterContext'
 import { DataContext } from '../../Context/DataContext'
 import { TeamCharacterContext } from '../../Context/TeamCharacterContext'
-import { nodeVStr, resolveInfo } from '../../Formula/uiData'
 const cols = {
   xs: 1,
   md: 2,
