+++ apps/frontend/src/app/PageTeam/CharacterDisplay/Tabs/TabOptimize/Components/ArtifactSetConfig.tsx
@@ -17,7 +17,7 @@ import {
 } from '@genshin-optimizer/gi/db'
 import { useDatabase, useOptConfig } from '@genshin-optimizer/gi/db-ui'
 import { SlotIcon } from '@genshin-optimizer/gi/svgicons'
-import { Translate } from '@genshin-optimizer/gi/ui'
+import { Translate, UIData } from '@genshin-optimizer/gi/ui'
 import { constant } from '@genshin-optimizer/gi/wr'
 import { CheckBox, CheckBoxOutlineBlank, Replay } from '@mui/icons-material'
 import BlockIcon from '@mui/icons-material/Block'
@@ -46,7 +46,6 @@ import { DataContext } from '../../../../../Context/DataContext'
 import { TeamCharacterContext } from '../../../../../Context/TeamCharacterContext'
 import { getArtSheet, setKeysByRarities } from '../../../../../Data/Artifacts'
 import { artifactDefIcon } from '../../../../../Data/Artifacts/ArtifactSheet'
-import { UIData } from '../../../../../Formula/uiData'
 import type { SetNum } from '../../../../../Types/consts'
 import { bulkCatTotal } from '../../../../../Util/totalUtils'
 import SetInclusionButton from './SetInclusionButton'
