+++ apps/frontend/src/app/PageTeam/CharacterDisplay/Tabs/TabOptimize/Components/OptimizationTargetSelector.tsx
@@ -2,12 +2,12 @@ import { useBoolState } from '@genshin-optimizer/common/react-util'
 import { ImgIcon, SqBadge } from '@genshin-optimizer/common/ui'
 import { objPathValue } from '@genshin-optimizer/common/util'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
+import { resolveInfo, type NodeDisplay } from '@genshin-optimizer/gi/ui'
 import { Box, Button, Divider, Stack } from '@mui/material'
 import { useCallback, useContext, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { DataContext } from '../../../../../Context/DataContext'
 import { getDisplayHeader } from '../../../../../Formula/DisplayUtil'
-import { resolveInfo, type NodeDisplay } from '../../../../../Formula/uiData'
 import type { TargetSelectorModalProps } from './TargetSelectorModal'
 import { TargetSelectorModal } from './TargetSelectorModal'
