+++ apps/frontend/src/app/PageTeam/CharacterDisplay/Tabs/TabOptimize/foreground.ts
@@ -8,17 +8,9 @@ import type { ICachedArtifact } from '@genshin-optimizer/gi/db'
 import type {
   ArtifactBuildData,
   ArtifactsBySlot,
-  DynStat,
 } from '@genshin-optimizer/gi/solver'
 import { getMainStatValue } from '@genshin-optimizer/gi/util'
-import type { Data, NumNode } from '@genshin-optimizer/gi/wr'
-import {
-  constant,
-  formulaString,
-  input,
-  setReadNodeKeys,
-} from '@genshin-optimizer/gi/wr'
-import { computeUIData } from '../../../../Formula/api'
+import { input, setReadNodeKeys } from '@genshin-optimizer/gi/wr'
 const dynamic = setReadNodeKeys(
   deepClone({ dyn: { ...input.art, ...input.artSet } })
 )
@@ -78,30 +70,30 @@ export function compactArtifacts(
   return result
 }
 
-export function debugCompute(
-  nodes: NumNode[],
-  base: DynStat,
-  arts: ArtifactBuildData[]
-) {
-  const stats = { ...base }
-  for (const art of arts) {
-    for (const [key, value] of Object.entries(art.values)) {
-      stats[key] = (stats[key] ?? 0) + value
-    }
-  }
-  const data = {
-    dyn: Object.fromEntries(
-      Object.entries(stats).map(([key, value]) => [key, constant(value)])
-    ),
-  } as Data
-  const uiData = computeUIData([data])
-  return {
-    base,
-    arts,
-    stats,
-    data,
-    uiData,
-    nodes: nodes.map(formulaString),
-    results: nodes.map((node) => uiData.get(node)),
-  }
-}
+// export function debugCompute(
+//   nodes: NumNode[],
+//   base: DynStat,
+//   arts: ArtifactBuildData[]
+// ) {
+//   const stats = { ...base }
+//   for (const art of arts) {
+//     for (const [key, value] of Object.entries(art.values)) {
+//       stats[key] = (stats[key] ?? 0) + value
+//     }
+//   }
+//   const data = {
+//     dyn: Object.fromEntries(
+//       Object.entries(stats).map(([key, value]) => [key, constant(value)])
+//     ),
+//   } as Data
+//   const uiData = computeUIData([data])
+//   return {
+//     base,
+//     arts,
+//     stats,
+//     data,
+//     uiData,
+//     nodes: nodes.map(formulaString),
+//     results: nodes.map((node) => uiData.get(node)),
+//   }
+// }
