+++ apps/frontend/src/app/PageTeam/CharacterDisplay/Tabs/TabTalent.tsx
@@ -3,6 +3,7 @@ import { range } from '@genshin-optimizer/common/util'
 import { maxConstellationCount } from '@genshin-optimizer/gi/consts'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
 import type { ICharacter } from '@genshin-optimizer/gi/good'
+import type { NodeDisplay } from '@genshin-optimizer/gi/ui'
 import { uiInput as input } from '@genshin-optimizer/gi/wr'
 import {
   Box,
@@ -28,7 +29,6 @@ import {
 import { CharacterContext } from '../../../Context/CharacterContext'
 import { DataContext } from '../../../Context/DataContext'
 import type { TalentSheetElementKey } from '../../../Data/Characters/ICharacterSheet'
-import type { NodeDisplay } from '../../../Formula/uiData'
 import type { DocumentSection } from '../../../Types/sheet'
 const talentSpacing = {
   xs: 12,
