+++ apps/frontend/src/app/PageTeam/TeamSetting/TeamComponents.tsx
@@ -14,6 +14,7 @@ import {
   useTeam,
   useTeamChar,
 } from '@genshin-optimizer/gi/db-ui'
+import type { NodeDisplay } from '@genshin-optimizer/gi/ui'
 import { ArtifactSetName } from '@genshin-optimizer/gi/ui'
 import { input } from '@genshin-optimizer/gi/wr'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
@@ -50,7 +51,6 @@ import { TeamCharacterContext } from '../../Context/TeamCharacterContext'
 import { dataSetEffects, getArtSheet } from '../../Data/Artifacts'
 import type CharacterSheet from '../../Data/Characters/CharacterSheet'
 import { resonanceSheets } from '../../Data/Resonance'
-import type { NodeDisplay } from '../../Formula/uiData'
 
 export function TeamBuffDisplay() {
   const { data, compareData } = useContext(DataContext)
