+++ apps/frontend/src/app/PageTeam/useCompareData.tsx
@@ -2,10 +2,10 @@ import { useForceUpdate } from '@genshin-optimizer/common/react-util'
 import { charKeyToLocCharKey } from '@genshin-optimizer/gi/consts'
 import type { ICachedArtifact, ICachedWeapon } from '@genshin-optimizer/gi/db'
 import { useDBMeta, useDatabase } from '@genshin-optimizer/gi/db-ui'
+import type { UIData } from '@genshin-optimizer/gi/ui'
 import type { Data } from '@genshin-optimizer/gi/wr'
 import { useContext, useDeferredValue, useEffect, useMemo } from 'react'
 import { TeamCharacterContext } from '../Context/TeamCharacterContext'
-import type { UIData } from '../Formula/uiData'
 import { getTeamDataCalc } from '../ReactHooks/useTeamData'
 import { getArtifactData } from './CharacterDisplay/Tabs/TabTheorycraft/optimizeTc'
