+++ apps/frontend/src/app/PageWeapon/WeaponCard.tsx
@@ -5,8 +5,9 @@ import {
 } from '@genshin-optimizer/common/ui'
 import { imgAssets, weaponAsset } from '@genshin-optimizer/gi/assets'
 import { useDatabase, useWeapon } from '@genshin-optimizer/gi/db-ui'
+import { computeUIData, nodeVStr, resolveInfo } from '@genshin-optimizer/gi/ui'
 import { ascensionMaxLevel } from '@genshin-optimizer/gi/util'
-import { uiInput as input } from '@genshin-optimizer/gi/wr'
+import { dataObjForWeapon, uiInput as input } from '@genshin-optimizer/gi/wr'
 import { Lock, LockOpen } from '@mui/icons-material'
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
 import EditIcon from '@mui/icons-material/Edit'
@@ -27,8 +28,6 @@ import { LocationAutocomplete } from '../Components/Character/LocationAutocomple
 import LocationName from '../Components/Character/LocationName'
 import type CharacterSheet from '../Data/Characters/CharacterSheet'
 import { getWeaponSheet } from '../Data/Weapons'
-import { computeUIData, dataObjForWeapon } from '../Formula/api'
-import { nodeVStr, resolveInfo } from '../Formula/uiData'
 import type { LocationKey } from '../Types/consts'
 
 type WeaponCardProps = {
