+++ apps/frontend/src/app/PageWeapon/WeaponEditor.tsx
@@ -3,8 +3,9 @@ import { StarsDisplay } from '@genshin-optimizer/common/ui'
 import { weaponAsset } from '@genshin-optimizer/gi/assets'
 import type { ICachedWeapon } from '@genshin-optimizer/gi/db'
 import { useDatabase, useWeapon } from '@genshin-optimizer/gi/db-ui'
+import { computeUIData } from '@genshin-optimizer/gi/ui'
 import { milestoneLevelsLow } from '@genshin-optimizer/gi/util'
-import { uiInput as input } from '@genshin-optimizer/gi/wr'
+import { dataObjForWeapon, uiInput as input } from '@genshin-optimizer/gi/wr'
 import { Lock, LockOpen } from '@mui/icons-material'
 import CloseIcon from '@mui/icons-material/Close'
 import {
@@ -31,7 +32,6 @@ import RefinementDropdown from '../Components/RefinementDropdown'
 import { DataContext } from '../Context/DataContext'
 import type CharacterSheet from '../Data/Characters/CharacterSheet'
 import { getWeaponSheet } from '../Data/Weapons'
-import { computeUIData, dataObjForWeapon } from '../Formula/api'
 import type { LocationKey } from '../Types/consts'
 const WeaponSelectionModal = React.lazy(
   () => import('../Components/Weapon/WeaponSelectionModal')
