+++ apps/frontend/src/app/ReactHooks/useTeamData.tsx
@@ -9,8 +9,15 @@ import {
   type ICachedWeapon,
 } from '@genshin-optimizer/gi/db'
 import { useDBMeta, useDatabase, useTeam } from '@genshin-optimizer/gi/db-ui'
-import type { Data } from '@genshin-optimizer/gi/wr'
-import { common } from '@genshin-optimizer/gi/wr'
+import { uiDataForTeam } from '@genshin-optimizer/gi/ui'
+import type { CharInfo, Data } from '@genshin-optimizer/gi/wr'
+import {
+  common,
+  dataObjForArtifact,
+  dataObjForCharacterNew,
+  dataObjForWeapon,
+  mergeData,
+} from '@genshin-optimizer/gi/wr'
 import { useContext, useDeferredValue, useEffect, useMemo } from 'react'
 import type { TeamData } from '../Context/DataContext'
 import { TeamCharacterContext } from '../Context/TeamCharacterContext'
@@ -20,14 +27,6 @@ import type CharacterSheet from '../Data/Characters/CharacterSheet'
 import { resonanceData } from '../Data/Resonance'
 import { getWeaponSheet } from '../Data/Weapons'
 import WeaponSheet from '../Data/Weapons/WeaponSheet'
-import type { CharInfo } from '../Formula/api'
-import {
-  dataObjForArtifact,
-  dataObjForCharacterNew,
-  dataObjForWeapon,
-  mergeData,
-  uiDataForTeam,
-} from '../Formula/api'
 import { getArtifactData } from '../PageTeam/CharacterDisplay/Tabs/TabTheorycraft/optimizeTc'
 
 type TeamDataBundle = {
