+++ apps/frontend/src/app/Types/consts.ts
@@ -4,48 +4,11 @@ import type {
   ElementKey,
   LocationCharacterKey,
 } from '@genshin-optimizer/gi/consts'
-import { absorbableEle } from '@genshin-optimizer/gi/consts'
 
 export const allArtifactSetCount = [1, 2, 3, 4, 5] as const
 
-/**
- * @deprecated use `allRollColorKeys` in `@genshin-optimizer/gi/ui`
- */
-export const allRollColorKeys = [
-  'roll1',
-  'roll2',
-  'roll3',
-  'roll4',
-  'roll5',
-  'roll6',
-] as const
-
-export const travelerFKeys = [
-  'TravelerAnemoF',
-  'TravelerGeoF',
-  'TravelerElectroF',
-  'TravelerDendroF',
-  'TravelerHydroF',
-] as const
-export const travelerMKeys = [
-  'TravelerAnemoM',
-  'TravelerGeoM',
-  'TravelerElectroM',
-  'TravelerDendroM',
-  'TravelerHydroM',
-] as const
-
 export type SetNum = (typeof allArtifactSetCount)[number]
 
-/**
- * @deprecated use `RollColorKey` in `@genshin-optimizer/gi/ui`
- */
-export type RollColorKey = (typeof allRollColorKeys)[number]
-
-/**
- * @deprecated use `absorbableEle` in `@genshin-optimizer/gi/consts`
- */
-export { absorbableEle }
 export const allowedAmpReactions: Dict<ElementKey, AmplifyingReactionKey[]> = {
   pyro: ['vaporize', 'melt'],
   hydro: ['vaporize'],
