+++ libs/gi/ui/src/index.ts
@@ -6,3 +6,4 @@ export * from './Theme'
 export * from './Translate'
 export * from './Weapon'
 export * from './getVariant'
+export * from './uiData'
