+++ libs/gi/ui/tsconfig.lib.json
@@ -4,7 +4,7 @@
     "outDir": "../../../dist/out-tsc",
     "types": [
       "node",
-
+      "object-overrides",
       "@nx/react/typings/cssmodule.d.ts",
       "@nx/react/typings/image.d.ts"
     ]
