+++ libs/gi/wr/src/index.ts
@@ -1,3 +1,4 @@
+export * from './api'
 export * from './debug'
 export * from './differentiate'
 export * from './formula'
